requirements.txt
Gradio==3.0.0
face_recognition==1.3.0
opencv-python==4.5.3.20210927
chromadb==0.3.21
numpy==1.21.2
Pillow==8.4.0
gradio==3.0.0
uuid==1.30
deepface==0.0.68